genx: enable option to exclude media kernels support
As of now option excludes media kernels support, but don't disable
CMRT path entirely, i.e. execution will fail on the attempt to load
the kernel though it is possible to fail earlier just starting
CMRT path. This is left for the future work.

Effect on Media SDK binary size (not stripped):
 * 41268416 before
 * 38263384 after

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>