vp9e: return VAAPI RT values as unit32_t
Accroding to VAAPI specification RT values  are defined as unit32_t.
For example:
  #define VA_RT_FORMAT_YUV420    0x00000001
  #define VA_RT_FORMAT_YUV420_10 0x00000100

In mediasdk code for vp9 encoder  we pretend that these values are
uint8_t which is wrong. This worked till we supported only those RT
formats which falls into 1st byte, but this will not work with,
for example, 10-bit YUV420. This patch fixes that.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>