[MDP-50760][VP9e] Fix rate control/frame rate buffer creation
rateParamBuf_ids starts being empty, but is resized to size of
at least 1 and initialized with 0 - this passes the VA_INVALID_ID
check, and the buffer with the ID 0 is destroyed (which is,
incidentally, the HRD buffer allocated just before the rate param
buffer); afterwards, a new buffer with the ID 0 is created and so on.
Obviously, this behavior is invalid - should resize the vector to
the new number of temporal layers only after the buffers with the IDs
previously contained in the vector are destroyed.