Return an error on warnings from SyncOperation
Currently the timeout used in SMT and sample_fei is 1500 seconds.
In regular case, it's even too much for task synchronization.
However in case of driver/MSDK synchronization problems,
SyncOperation may return MFX_WRN_IN_EXECUTION which was ignored.
It leaded to undefined behavior hard to debug.
So it's better to catch synchronization issues at once.

Signed-off-by: Dmitry Ermilov <dmitry.ermilov@intel.com>