build: allow build against system installed gtest with USE_SYSTEM_GTEST
Fixes: #1285

This change allows builds against system installed gtest if
cmake -DUSE_SYSTEM_GTEST=ON is specified. Mind that this change
does not provide any configure level check over gtest. And it does not
provide proper target configuration if gtest is installed to custom
location. Build will work if gtest will happen to be in the right place
at the right time.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>