Make sure to compare guids correctly
Fix for #1398

On Ubuntu 19.04 compiler (gcc-8.3.0) aligns GUID stucture size which
we defined internally in mediasdk by 8 and does not zero out the
structure when allocated on stack. As a result when we use current
approach to compare guids via !memcpy(a,b,sizeof(GUID)) we actually
compare unitialized regions and mistakenly think some guids not
equal which leads to runtime errors. For example, we might fail
to enable LowPower encoding on this command line:
  sample_encode h264 -w 176 -h 144 -f 30 -cqp \
    -qpi 30 -qpp 30 -qpb 30 -qsv-ff -nv12 -i a.yuv -o a.264

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>